+++ src/utils/common.js
@@ -3,8 +3,6 @@ export const humanizeCommentDate = (date) => dayjs(date).format('YYYY/MM/DD HH:m
 export const humanizeFilmReleaseDetailedDate = (date) => dayjs(date).format('DD MMMM YYYY');
 export const humanizeFilmReleaseDate = (date) => dayjs(date).format('YYYY');
 
-// Функция из интернета по генерации случайного числа из диапазона
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
